+++ src/full/Agda/TypeChecking/Monad/Base.hs
@@ -5999,6 +5999,7 @@ data TypeError
         | NeedOptionCubical Cubical String
             -- ^ Flavor of cubical needed for the given reason.
         | NeedOptionPatternMatching
+        | NeedOptionIrrelevance
         | NeedOptionProp
         | NeedOptionRewriting
         | NeedOptionSizedTypes String